+++ app/transports/socketio/routes/document.js
@@ -1,0 +1,19 @@
+// Utils
+const debug = require('debug')('canvas-transport-socketio-route-document')
+
+/**
+ * Constants
+ */
+
+const DOC_GET_BY_ID = 'document:get:id';
+
+
+/**
+ * Context routes
+ * @param {*} socket
+ * @param {*} context
+ */
+
+module.exports = function (socket, context) {
+
+};
