+++ bin/update-repos.sh
@@ -2,6 +2,7 @@
 
 
 # Creadits for this script goes to chatgpt
+# Do not use, will update it later
 
 
 # Find the Canvas directory from the current path
@@ -23,7 +24,7 @@ update_repos() {
             echo "Updating repository: $(basename "$dir")"
             cd "$dir"
             git fetch --all
-            git pull
+
             if [ $? -ne 0 ]; then
                 echo "Merge conflict in $(basename "$dir"). Stopping script."
                 exit 1
